+++ packages/alias/CHANGELOG.md
@@ -1,5 +1,12 @@
 # xplainer
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [e013f7a]
+  - @xplainer/cli@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes